+++ backend/canisters/group/impl/src/updates/c2c_tip_message.rs
@@ -35,7 +35,11 @@ fn c2c_tip_message_impl(args: Args, state: &mut RuntimeState) -> Response {
         now,
     };
 
-    match state.data.chat.tip_message(tip_message_args) {
+    match state
+        .data
+        .chat
+        .tip_message(tip_message_args, &mut state.data.event_sink_client)
+    {
         TipMessageResult::Success => {
             if let Some((message_index, message_event_index)) = state
                 .data