+++ backend/libraries/chat_events/src/chat_events.rs
@@ -18,16 +18,16 @@ use types::{
     CompletedCryptoTransaction, Cryptocurrency, DirectChatCreated, EventIndex, EventWrapper, EventWrapperInternal,
     EventsTimeToLiveUpdated, GroupCanisterThreadDetails, GroupCreated, GroupFrozen, GroupUnfrozen, Hash, HydratedMention,
     Mention, Message, MessageContentInitial, MessageEventPayload, MessageId, MessageIndex, MessageMatch, MessageReport,
-    Milliseconds, MultiUserChat, P2PSwapAccepted, P2PSwapCompletedEventPayload, P2PSwapContent, P2PSwapStatus,
-    PendingCryptoTransaction, PollVotes, ProposalUpdate, PushEventResult, Reaction, RegisterVoteResult, ReserveP2PSwapResult,
-    ReserveP2PSwapSuccess, TimestampMillis, TimestampNanos, Timestamped, Tips, UserId, VideoCall, VoteOperation,
+    MessageTippedEventPayload, Milliseconds, MultiUserChat, P2PSwapAccepted, P2PSwapCompletedEventPayload, P2PSwapContent,
+    P2PSwapStatus, PendingCryptoTransaction, PollVotes, ProposalUpdate, PushEventResult, Reaction, RegisterVoteResult,
+    ReserveP2PSwapResult, ReserveP2PSwapSuccess, TimestampMillis, TimestampNanos, Timestamped, Tips, UserId, VideoCall,
+    VoteOperation,
 };
 
 pub const OPENCHAT_BOT_USER_ID: UserId = UserId::new(Principal::from_slice(&[228, 104, 142, 9, 133, 211, 135, 217, 129, 1]));
 
 #[derive(Serialize, Deserialize)]
 pub struct ChatEvents {
-    #[serde(default = "default_chat")]
     chat: Chat,
     main: ChatEventsList,
     threads: HashMap<MessageIndex, ChatEventsList>,
@@ -38,28 +38,10 @@ pub struct ChatEvents {
     expiring_events: ExpiringEvents,
     last_updated_timestamps: LastUpdatedTimestamps,
     pub video_call_in_progress: Timestamped<Option<VideoCall>>,
-    #[serde(default)]
     anonymized_id: String,
 }
 
-fn default_chat() -> Chat {
-    Chat::Direct(Principal::anonymous().into())
-}
-
 impl ChatEvents {
-    // TODO POST RELEASE - remove this
-    pub fn set_chat(&mut self, chat: Chat) {
-        self.chat = chat;
-    }
-
-    pub fn anonymized_id(&self) -> String {
-        self.anonymized_id.clone()
-    }
-
-    pub fn set_anonymized_id(&mut self, id: u128) {
-        self.anonymized_id = hex::encode(id.to_be_bytes());
-    }
-
     pub fn new_direct_chat(
         them: UserId,
         events_ttl: Option<Milliseconds>,
@@ -120,6 +102,10 @@ impl ChatEvents {
         );
 
         events
+    }
+
+    pub fn set_chat(&mut self, chat: Chat) {
+        self.chat = chat;
     }
 
     pub fn iter_recently_updated_events(
@@ -643,7 +629,12 @@ impl ChatEvents {
         }
     }
 
-    pub fn tip_message(&mut self, args: TipMessageArgs, min_visible_event_index: EventIndex) -> TipMessageResult {
+    pub fn tip_message<R: Runtime + Send + 'static>(
+        &mut self,
+        args: TipMessageArgs,
+        min_visible_event_index: EventIndex,
+        event_sink_client: Option<&mut EventSinkClient<R>>,
+    ) -> TipMessageResult {
         use TipMessageResult::*;
 
         if let Some((message, event_index)) = self.message_internal_mut(
@@ -660,8 +651,25 @@ impl ChatEvents {
 
             message.tips.push(args.ledger, args.user_id, args.amount);
             message.last_updated = Some(args.now);
-            self.last_updated_timestamps
-                .mark_updated(args.thread_root_message_index, event_index, args.now);
+
+            if let Some(client) = event_sink_client {
+                let message_type = message.content.message_type();
+
+                client.push(
+                    EventBuilder::new("message_tipped", args.now)
+                        .with_user(args.user_id.to_string())
+                        .with_source(self.chat.canister_id().to_string())
+                        .with_json_payload(&MessageTippedEventPayload {
+                            message_type,
+                            chat_type: self.chat.chat_type().to_string(),
+                            chat_id: self.anonymized_id.clone(),
+                            thread: args.thread_root_message_index.is_some(),
+                            token: args.token.token_symbol().to_string(),
+                            amount: args.amount,
+                        })
+                        .build(),
+                );
+            }
 
             add_to_metrics(
                 &mut self.metrics,
@@ -670,6 +678,9 @@ impl ChatEvents {
                 |m| incr(&mut m.tips),
                 args.now,
             );
+
+            self.last_updated_timestamps
+                .mark_updated(args.thread_root_message_index, event_index, args.now);
 
             Success
         } else {